Add files via upload
Temperature Converter
This is a simple web-based temperature converter that allows users to convert temperatures between Celsius, Fahrenheit, and Kelvin.

Features
Convert temperatures between Celsius, Fahrenheit, and Kelvin.
Easy-to-use interface with input fields and dropdown menus.
Responsive design for seamless usage on various devices.
Getting Started
To use the temperature converter, simply open the index.html file in your web browser.

Usage
Enter the temperature you want to convert in the input field labeled "Enter Temperature".
Select the unit you want to convert from in the dropdown menu labeled "Convert From".
Select the unit you want to convert to in the dropdown menu labeled "Convert To".
Click the "Convert" button to see the converted temperature in the designated unit.
About
This temperature converter was created as a simple project to demonstrate HTML, CSS, and JavaScript skills. It utilizes basic web development techniques to provide a functional tool for converting temperatures.